pyproject.toml
[tool.poetry]
name = "vm-uncovered"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vm_uncovered"}]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.3.2"
python-dotenv = "^0.21.0"
motor = "^3.1.1"
starlite = "^1.35.1"
uvicorn = "^0.19.0"
Jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
mypy = "^0.982"
pytest = "^7.2.0"
mongo-types = "^0.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 109
target-version = ['py10']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | *_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''

[tool.isort]
profile = "black"
line_length = 109

[tool.mypy]
python_version = 3.10
allow_untyped_defs = true
follow_imports = "silent"
ignore_missing_imports = true
show_column_numbers = true
pretty = true

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = ["."]
addopts = "-ra -q"
testpaths = ["tests"]